requirements.txt
GitPython==0.3.6
arrow==0.5.4
gitdb==0.6.4
python-dateutil==2.4.1
six==1.9.0
smmap==0.9.0
